+++ package.json
@@ -26,6 +26,10 @@
     "type": "git",
     "url": "https://github.com/ember-cli/ember-cli.git"
   },
+  "@std/esm": {
+    "esm": "js",
+    "cjs": true
+  },
   "scripts": {
     "docs": "yuidoc",
     "lint": "node tests/runner lint",
@@ -37,6 +41,7 @@
     "test:debug": "node debug tests/runner"
   },
   "dependencies": {
+    "@std/esm": "^0.9.0",
     "amd-name-resolver": "1.0.0",
     "babel-plugin-transform-es2015-modules-amd": "^6.24.0",
     "bower-config": "^1.3.0",
